Closes #102 - drops old Edge support & switched to a pure dynamic runtime
