+++ docs/guides/adding-npm-packages.md
@@ -1,7 +1,7 @@
 ## Working With NPM Packages
 
-Sometimes you'll need to work with NPM packages and 3rd party libraries. The easiest way to consume NPM packages is to use a tool like [Snowpack](https://www.snowpack.dev/). Snowpack unpacks packages as single JavaScript files in a `web_modules/` directory and configures the package to be imported using ES Modules.
+Sometimes you'll need to work with NPM packages and 3rd party libraries. The easiest way to consume NPM packages is to use a tool like [Snowpack](https://www.snowpack.dev/). Snowpack unpacks packages as a single JavaScript file configures the package to be imported using ES Modules.
 
 ## 3rd Party Libraries
 
-Not all JavaScript libraries are available via NPM and not all NPM packages can be exported as ES Modules. When this happens simply include the library via the `fetchJS` function.
+Not all JavaScript libraries are available via NPM and not all NPM packages can be exported as ES Modules.
