+++ src/core/body-parser.ts
@@ -1,4 +1,4 @@
-import { fetchCSS } from "./fetch";
+import { djinnjsOutDir } from "./config";
 
 export async function parse(el: HTMLElement) {
     let files = [];
@@ -7,6 +7,9 @@ export async function parse(el: HTMLElement) {
         const strings = attr.split(/\s+/g);
         files = [...files, ...strings];
     });
-    await fetchCSS(files);
+    if (files.length) {
+        const { fetchCSS } = await import(`${location.origin}/${djinnjsOutDir}/fetch.mjs`);
+        await fetchCSS(files);
+    }
     return;
 }
